Do not remove empty clauses of a pattern (#18)
* Do not remove empty clauses of a pattern

The FlattenPatterns pass expects pattern matches to pass
scilla-checker's PatternChecker. Doing DCE on it will lead
to redundant patterns, invalidating the premise for FlattenPatterns.

Such DCE for empty match branches can be done at a later stage,
where the pattern-match itself may possibly be eliminated. It has
a broader scope, and is hence for the future.

Add a unit test for this, and the ud-registry contract (which
was the one where I actually faced this issue).

* Update gold file for codegen contract test