Create explicit variable declaration (allocation) statements.
While the renaming pass prevents name clashes (shadowing) of variables,
it doesn't make the variables truly global because we can still have
two variables with the same name defined in different clauses of a match,
and with possibly different types. This indicates that it's simpler
to instead declare variables in their scope before definition. Since
the renamer still runs, we cannot end up with two declarations of
different types but with the same name.

This commit also fixes a bug in the ScopingRename pass where definitions
inside the LHS of a bind may be shadowed by the variable defined in the
bind itself.

for example, in the below expression, we don't want the inner `a` to
be shadowed by the outer `a` after statement flattening (in closure
conversion), and hence we rename in this case also.

```
  let a =
      let a = Uint32 0 in
      let b = builtin to_uint64 a in
      b
  in
  a
```