Fix names in type value_scope.
The name "Global" was used to represent function declarations
in the environment. These values don't need to be loaded
(call instructions take pointer to functions as argument).
However actual global values will need to be loaded. This
creates a confusion in resolve_id_value. Hence rename
"Global" to "FunDecl", which is what it actually is currently.

Create a new name "Global" to correspond to llvm::GlobalVariable.

Unrelated changes packed with this commit include wrapping
Llvm.build_insertvalue and Llvm.build_extractvalue to check
the right type and field index. There's also a large comment
on how (as we progress) paramters/return values with builtins
(functions) in SRTL will be passed around, especially when
multiple types are possible.