First version of (compile time) libeval
This version cannot handle closures, and if there's a closure
in the definition of a library variable, it falls back to
having the original definition rather than a (partially)
evaluated one.

This one isn't good enough for us though, it cannot eliminate
unused type applications (such as `int32_neq`) we get from IntUtils.